# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FocusBlock API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Task {
  id: number;
  name: string;
  date: string;
  /** 1-based index within its block group */
  chunkIndex: number;
  totalChunks: number;
  completed: boolean;
  /** Number of times the timer was triggered for this task group */
  plays: number;
  /**
     * Task-group note / plan (stored on the first block of the group)
     * @nullable
     */
  description?: string | null;
  /**
     * Reference URL attached to the task group
     * @nullable
     */
  link?: string | null;
  /**
     * Emoji rating: 😩 | 😐 | 🙂 | 🔥
     * @nullable
     */
  rating?: string | null;
  /**
     * What was worked on during this block
     * @nullable
     */
  notes?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface TaskInput {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 16
     */
  chunks: number;
  date?: string;
}

export type TaskUpdateRating = typeof TaskUpdateRating[keyof typeof TaskUpdateRating];


export const TaskUpdateRating = {
  '😩': '😩',
  '😐': '😐',
  '🙂': '🙂',
  '🔥': '🔥',
} as const;

export interface TaskUpdate {
  completed?: boolean;
  name?: string;
  rating?: TaskUpdateRating;
  /** What was worked on during this block */
  notes?: string;
  /** Task-group note / plan (empty string clears it) */
  description?: string;
  /** Reference URL for the task group (empty string clears it) */
  link?: string;
  /** @minimum 0 */
  plays?: number;
  /**
     * Atomically add this many plays to the current value
     * @minimum 1
     */
  incrementPlays?: number;
}

export interface CompleteFocusInput {
  /** @minLength 1 */
  name: string;
}

export interface TaskNote {
  /**
     * Id of the block holding the note (null when no group exists yet)
     * @nullable
     */
  taskId: number | null;
  /**
     * The task-group note text (null when empty)
     * @nullable
     */
  description: string | null;
}

export interface TaskNoteInput {
  /** @minLength 1 */
  name: string;
  /** Note text (empty string clears the note) */
  description: string;
}

export interface TaskGroupContext {
  name: string;
  /**
     * Calendar day of the resolved group (null when no group exists yet)
     * @nullable
     */
  date?: string | null;
  blocks: Task[];
  /** Originally planned block count for the group */
  totalEstimated: number;
  /** Number of completed blocks (may exceed totalEstimated when over) */
  completedCount: number;
  /**
     * The task-group note / plan
     * @nullable
     */
  description?: string | null;
  /**
     * Reference URL attached to the task group
     * @nullable
     */
  link?: string | null;
}

export interface AddBlockInput {
  /** @minLength 1 */
  name: string;
}

export interface TimelinePage {
  tasks: Task[];
  /**
     * Pass as `before` to load the next (older) page; null when no older days exist
     * @nullable
     */
  nextCursor?: string | null;
  hasMore: boolean;
}

export interface TodayStats {
  totalBlocks: number;
  totalMinutes: number;
  /** Remaining planned focus blocks across all open task groups (any day), clamped per group so over-estimated groups never go negative. */
  openPlannedBlocks: number;
}

export type GetTimelineParams = {
/**
 * Only return days strictly before this calendar date (YYYY-MM-DD)
 */
before?: string;
/**
 * Number of distinct days to return in this page
 * @minimum 1
 * @maximum 31
 */
limit?: number;
};

export type GetTaskNoteParams = {
/**
 * Task name to look up
 * @minLength 1
 */
name: string;
};

export type GetTaskGroupParams = {
/**
 * Task name to look up
 * @minLength 1
 */
name: string;
};