# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FocusBlock API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Focus block / timeline operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    post:
      operationId: createTasks
      tags: [tasks]
      summary: Create planner tasks (expands chunks)
      description: Creates one task row per chunk (e.g. 3 chunks creates 3 rows with the same name)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Tasks created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  /tasks/timeline:
    get:
      operationId: getTimeline
      tags: [tasks]
      summary: Paginated timeline of tasks across days
      description: >-
        Returns tasks grouped-ready for the timeline, newest day first, a page
        of days at a time. Pass the previous page's nextCursor as `before` to
        load older days.
      parameters:
        - name: before
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Only return days strictly before this calendar date (YYYY-MM-DD)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 31
            default: 7
          description: Number of distinct days to return in this page
      responses:
        "200":
          description: A page of timeline tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelinePage"

  /tasks/stats/today:
    get:
      operationId: getTodayStats
      tags: [tasks]
      summary: Get today's focus stats
      description: Returns aggregated stats for today (completed blocks and minutes focused)
      responses:
        "200":
          description: Today's stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayStats"

  /tasks/complete-focus:
    post:
      operationId: completeFocus
      tags: [tasks]
      summary: Complete a focus block for a task by name
      description: >-
        Marks the next open block for the given name (today) as done. If no open
        block exists, a new completed block is created for today. Returns the
        completed task so the caller can attach a rating.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteFocusInput"
      responses:
        "200":
          description: Block completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/note:
    get:
      operationId: getTaskNote
      tags: [tasks]
      summary: Get a task's note by name
      description: >-
        Returns the task-group note (description) for the most recent group with
        the given name (any day), or null if no such group exists yet.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Task name to look up
      responses:
        "200":
          description: The task's note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskNote"
    put:
      operationId: updateTaskNote
      tags: [tasks]
      summary: Set a task's note by name
      description: >-
        Upserts the task-group note for the most recent group with the given
        name (any day). Updates that group's first block if it exists; otherwise
        creates a single open block today to hold the note (skipped when the note
        is empty). An empty note clears an existing one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskNoteInput"
      responses:
        "200":
          description: The updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskNote"

  /tasks/group:
    get:
      operationId: getTaskGroup
      tags: [tasks]
      summary: Get a task group's context by name
      description: >-
        Returns the active block group for the given name — today's group if it
        exists, otherwise the most recent group (any day). Includes the group's
        blocks, the estimated vs. completed counts, and the note/link. Blocks are
        empty when no group exists yet for the name.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Task name to look up
      responses:
        "200":
          description: The task group's context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskGroupContext"

  /tasks/add-block:
    post:
      operationId: addBlock
      tags: [tasks]
      summary: Append an extra open block to a task group
      description: >-
        Adds one more open block to today's group for the given name without
        changing the estimate, so completing it beyond the plan shows up as
        overage. Returns the created block.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBlockInput"
      responses:
        "200":
          description: Block added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      description: Mark a task done, rename it, attach a rating/notes, or adjust plays
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      required: [id, name, date, chunkIndex, totalChunks, completed, plays, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        date:
          type: string
          format: date
        chunkIndex:
          type: integer
          description: 1-based index within its block group
        totalChunks:
          type: integer
        completed:
          type: boolean
        plays:
          type: integer
          description: Number of times the timer was triggered for this task group
        description:
          type: ["string", "null"]
          description: "Task-group note / plan (stored on the first block of the group)"
        link:
          type: ["string", "null"]
          description: "Reference URL attached to the task group"
        rating:
          type: ["string", "null"]
          description: "Emoji rating: 😩 | 😐 | 🙂 | 🔥"
        notes:
          type: ["string", "null"]
          description: "What was worked on during this block"
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time

    TaskInput:
      type: object
      required: [name, chunks]
      properties:
        name:
          type: string
          minLength: 1
        chunks:
          type: integer
          minimum: 1
          maximum: 16
        date:
          type: string
          format: date

    TaskUpdate:
      type: object
      properties:
        completed:
          type: boolean
        name:
          type: string
        rating:
          type: string
          enum: ["😩", "😐", "🙂", "🔥"]
        notes:
          type: string
          description: "What was worked on during this block"
        description:
          type: string
          description: "Task-group note / plan (empty string clears it)"
        link:
          type: string
          description: "Reference URL for the task group (empty string clears it)"
        plays:
          type: integer
          minimum: 0
        incrementPlays:
          type: integer
          minimum: 1
          description: Atomically add this many plays to the current value

    CompleteFocusInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    TaskNote:
      type: object
      required: [taskId, description]
      properties:
        taskId:
          type: ["integer", "null"]
          description: Id of the block holding the note (null when no group exists yet)
        description:
          type: ["string", "null"]
          description: The task-group note text (null when empty)

    TaskNoteInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          description: Note text (empty string clears the note)

    TaskGroupContext:
      type: object
      required: [name, blocks, totalEstimated, completedCount]
      properties:
        name:
          type: string
        date:
          type: ["string", "null"]
          format: date
          description: Calendar day of the resolved group (null when no group exists yet)
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        totalEstimated:
          type: integer
          description: Originally planned block count for the group
        completedCount:
          type: integer
          description: Number of completed blocks (may exceed totalEstimated when over)
        description:
          type: ["string", "null"]
          description: The task-group note / plan
        link:
          type: ["string", "null"]
          description: Reference URL attached to the task group

    AddBlockInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    TimelinePage:
      type: object
      required: [tasks, hasMore]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        nextCursor:
          type: ["string", "null"]
          format: date
          description: Pass as `before` to load the next (older) page; null when no older days exist
        hasMore:
          type: boolean

    TodayStats:
      type: object
      required: [totalBlocks, totalMinutes, openPlannedBlocks]
      properties:
        totalBlocks:
          type: integer
        totalMinutes:
          type: integer
        openPlannedBlocks:
          type: integer
          description: Remaining planned focus blocks across all open task groups (any day), clamped per group so over-estimated groups never go negative.