# lib/api-zod/src/generated/types/addBlockInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FocusBlock API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddBlockInput {
  /** @minLength 1 */
  name: string;
}