# lib/api-zod/src/generated/types/taskGroupContext.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FocusBlock API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from './task';

export interface TaskGroupContext {
  name: string;
  /**
     * Calendar day of the resolved group (null when no group exists yet)
     * @nullable
     */
  date?: Date | null;
  blocks: Task[];
  /** Originally planned block count for the group */
  totalEstimated: number;
  /** Number of completed blocks (may exceed totalEstimated when over) */
  completedCount: number;
  /**
     * The task-group note / plan
     * @nullable
     */
  description?: string | null;
  /**
     * Reference URL attached to the task group
     * @nullable
     */
  link?: string | null;
}